requirements.txt
django-suit==0.3a1
django-smartfields
pillow
